Additional functionality for PlayingActivity.
Added buttons for play,pause,next song and previous song. Removed "Now
Playing" textview because it seemed unnecessary. Modified getSongs
function, now it also gets album art bitmap and album name. Fixed a bug
with seekBar in playing activity, if you were to hold it would start to
go to song current position and back to where we were holding it which
looked a bit weird.